scripts/requirements.txt
feedparser==6.0.11
requests==2.31.0
python-dateutil==2.8.2
beautifulsoup4==4.12.3
lxml==5.1.0
